+++ website/package.json
@@ -19,6 +19,7 @@
     "@docusaurus/core": "2.0.0-beta.18",
     "@docusaurus/preset-classic": "2.0.0-beta.18",
     "@mdx-js/react": "^1.6.22",
+    "@types/react-loadable": "^5.5.6",
     "clsx": "^1.1.1",
     "docusaurus-lunr-search": "^2.1.15",
     "firebase": "^9.6.10",
@@ -27,6 +28,7 @@
     "prism-react-renderer": "^1.3.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-loadable": "^5.5.0",
     "suspend-react": "^0.0.8",
     "worker-loader": "^3.0.8"
   },
